+++ src/components/Card/types.ts
@@ -1,0 +1,7 @@
+export const enum CardComponentType {
+  knowledgeBase = 'knowledge-base',
+  problemSolver = 'problem-solver',
+  interface = 'interface',
+  subSystem = 'subsystem',
+  unknown = 'unknown',
+}