+++ src/components/CardInfo/types.tsx
@@ -1,0 +1,4 @@
+export enum InstallMethodType {
+  DynamicallyInstalledComponent = 'dynamicallyInstalledComponent',
+  ReusableComponent = 'reusableComponent',
+}