+++ src/components/Router/Router.tsx
@@ -4,6 +4,7 @@ import { Layout } from '@components/Layout';
 import { routes, DEFAULT_COMMAND_PATH } from '@constants';
 import { Action } from '@pages/Action';
 import { Command } from '@pages/Command';
+import { Library } from '@pages/Library';
 import { Main } from '@pages/Main';
 
 export const Router = () => {
@@ -21,6 +22,7 @@ export const Router = () => {
           <Route index element={<Navigate to={DEFAULT_COMMAND_PATH} replace />} />
           <Route path={routes.COMMAND} element={<Command />} />
           <Route path={routes.ACTION} element={<Action />} />
+          <Route path={routes.LIBRARY} element={<Library />} />
         </Route>
       </Routes>
     </Suspense>