+++ src/components/SidePanel/SidePanel.tsx
@@ -14,6 +14,7 @@ import { useSelector } from '@hooks';
 import { selectUser } from '@store/commonSlice';
 import { selectRequests, setRequests } from '@store/requestHistorySlice';
 import styles from './SidePanel.module.scss';
+import { SwitchMode } from './SwitchMode';
 
 interface IProps {
   className?: string;
@@ -56,6 +57,8 @@ export const SidePanel: FC<IProps> = ({ className }) => {
             [styles.accordionContent_admin]: !!user?.is_admin,
           })}
         >
+          <SwitchMode />
+
           <div>
             <Accordion
               header={translate({ ru: 'Разделы', en: 'Sections' })}